\name{extree}
\alias{extree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Extensible Trees
}
\description{
  Extensible trees provide the basic infrastructure to define
  tree algorithms via transformation, variable selection, and
  split point selection functions.
}
\usage{
extree(data, trafo, control = extree_control(...), 
    converged = NULL, ...)
}
\arguments{
  \item{data}{an object of class \code{\link{extree_data}}, see
\code{\link{extree_data}}.
}
  \item{trafo}{a function with arguments \code{subset}, \code{data},
	  \code{weights}, \code{info}, \code{estfun} and \code{object}.
}
  \item{converged}{an optional function with arguments \code{subset},
	  \code{weights}.
}
  \item{control}{list of control arguments generated by
	  \code{\link{extree_control}}.  Among others the variable and split
	  point selection (\code{selectfun}, \code{splitfun}) are specified
	  here.
}
  \item{...}{Additional arguments passed on to \code{extree_fit}.
}
}
\details{
This basic tree algorithm can be used to define your own tree algorithm 
variants. 

\code{trafo} defines how you want to preprocess you data for variable and split
point selection.  As an example, \code{mob} computes a model and returns
information such as \code{estfun} (the empirical estimating functions / score
contribution matrix, see also \code{\link[sandwich]{estfun}}), \code{objfun}
(value of the minimized objective function, usually negative log-Likelihood),
\code{coef} (estimated model coefficients), and \code{converged} (logical, has
the model converged?).

\code{selectfun} defines how to select the split variable.
\code{splitfun} defines how to select the split point. 
Details in \code{\link{extree_control}}
}
\value{Currently: A list of \code{nodes} (an object of class \code{partynode})
	and \code{trafo} (the encapsulated transformation function).
	This will likely change soon.
}
\keyword{tree}
